requirements.txt
streamlit
pandas
plotly
pymavlink
google-generativeai
numpy